Implement offset specifier to streaming API
Fixes: #79.

Streaming API now allows for "Streaming Offset" to be passed in.

Each event in the streaming API has it's own offset value. The offset value is
assured to increase as the stream progresses. Users of the streaming api are
advised to keep track of the offset value of the last event they processed.
The offset value can then be used to skip the the seen events, should the stream
disconnect.