Provide explicit name for Falcon Cloud Region autodiscovery
This will allow us to include the following excerpt to documentation:

    FALCON_CLOUD="autodiscover"

Users seeing this, will know the option exists, while at the same time they will
be able to get away with not setting it. Sharing the information about the
option existence is beneficial to certain class of users, who like to be
particularly explicit about stuff.